pyproject.toml
[project]
name = "image-ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "pillow>=12.1.0",
    "pytesseract>=0.3.13",
    "python-multipart>=0.0.21",
]
